test(libsyntheticgen,msbridge): split parser and msbridge by behaviour family
- parser.test.js (425) → parser-structure.test.js (terrain/org/
  department-team/people sections) and parser-content.test.js (project/
  snapshots/scenario/content/default-AST/error-handling); shared parseDsl
  reproduced per sibling.
- msbridge.test.js (442) → msbridge-contract.test.js (module exports,
  validateCallbackPayload, construction) and msbridge-runtime.test.js
  (callback handler, ConversationReference round-trip, acknowledgement
  lifecycle); shared makeConfig/makeAdapter/makeGhuserClient/newService
  lifted to msbridge-helpers.js.

Assertions unchanged; test counts preserved (32, 15).
Spec 0640 § C (large files), design Decision 5.
