+++ services/msbridge/test/msbridge-runtime.test.js
@@ -1,190 +1,10 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
-import {
-  createMockConfig,
-  createMockDiscussionClient,
-  createMockLogger,
-  createMockTracer,
-  createMockClock,
-} from "@forwardimpact/libmock";
+import { createMockLogger } from "@forwardimpact/libmock";
 
-import {
-  MsBridgeService,
-  appendHistory,
-  buildPrompt,
-  validateCallbackPayload,
-} from "../index.js";
-import {
-  DEFAULT_TICKET_SECRET,
-  DEFAULT_TRUSTED_ORIGINS,
-  createStatefulDiscussionClient,
-} from "./helpers.js";
-
-function makeConfig(overrides = {}) {
-  return createMockConfig("msbridge", {
-    port: 0,
-    host: "127.0.0.1",
-    github_repo: "owner/repo",
-    callback_base_url: "https://tunnel.example",
-    msAppId: () => "test-app-id",
-    msAppPassword: () => "test-password",
-    msAppTenantId: () => "test-tenant",
-    ...overrides,
-  });
-}
-
-function makeAdapter(overrides = {}) {
-  const sent = [];
-  const reactionActivities = [];
-  return {
-    sent,
-    reactionActivities,
-    process: async (_req, res, callback) => {
-      const turnContext = {
-        activity: overrides.activity ?? { type: "message" },
-        sendActivity: async (activity) => {
-          sent.push(activity);
-        },
-      };
-      await callback(turnContext);
-      if (!res.headersSent) res.status(200).end();
-    },
-    continueConversationAsync: async (_appId, _ref, callback) => {
-      await callback({
-        sendActivity: async (activity) => {
-          if (
-            activity &&
-            typeof activity === "object" &&
-            activity.type === "messageReaction"
-          ) {
-            reactionActivities.push(activity);
-          } else {
-            sent.push(activity);
-          }
-        },
-      });
-    },
-    onTurnError: null,
-    ...overrides,
-  };
-}
-
-function makeGhuserClient(token = "ghs_per_user") {
-  return {
-    GetToken: async () => ({ result: "token", token }),
-  };
-}
-
-function newService({
-  adapter,
-  config: configOverrides,
-  logger,
-  ghuserClient,
-} = {}) {
-  return new MsBridgeService(makeConfig(configOverrides), {
-    logger: logger ?? createMockLogger(),
-    tracer: createMockTracer(),
-    clock: createMockClock(),
-    discussionClient: createStatefulDiscussionClient(),
-    ghuserClient: ghuserClient ?? makeGhuserClient(),
-    adapter: adapter ?? makeAdapter(),
-    trustedOrigins: DEFAULT_TRUSTED_ORIGINS,
-    ticketSecret: DEFAULT_TICKET_SECRET,
-  });
-}
+import { newService, makeAdapter } from "./msbridge-helpers.js";
 
 describe("msbridge service", () => {
-  describe("module exports", () => {
-    test("exports MsBridgeService class", () => {
-      expect(typeof MsBridgeService).toBe("function");
-      expect(MsBridgeService.prototype.start).toBeTruthy();
-      expect(MsBridgeService.prototype.stop).toBeTruthy();
-    });
-
-    test("re-exports buildPrompt, appendHistory, validateCallbackPayload from libbridge", () => {
-      expect(typeof buildPrompt).toBe("function");
-      expect(typeof appendHistory).toBe("function");
-      expect(typeof validateCallbackPayload).toBe("function");
-    });
-  });
-
-  describe("validateCallbackPayload (lenient libbridge contract)", () => {
-    test("requires only correlation_id", () => {
-      expect(validateCallbackPayload(null)).toBeNull();
-      expect(validateCallbackPayload({})).toBeNull();
-      const minimal = validateCallbackPayload({ correlation_id: "c-1" });
-      expect(minimal).toEqual({
-        correlation_id: "c-1",
-        kind: "terminal",
-        seq: -1,
-        body: "",
-        agent: "",
-        last_acted_seq: -1,
-        verdict: "unknown",
-        summary: "",
-        replies: [],
-      });
-    });
-
-    test("passes through optional channel-agnostic fields", () => {
-      const payload = validateCallbackPayload({
-        correlation_id: "c-1",
-        verdict: "adjourned",
-        summary: "done",
-        replies: [{ body: "hi" }],
-        trigger: { kind: "missing_input", replies: 2 },
-        discussion_id: "GD_x",
-      });
-      expect(payload.replies).toEqual([{ body: "hi" }]);
-      expect(payload.trigger).toEqual({ kind: "missing_input", replies: 2 });
-      expect(payload.discussion_id).toBe("GD_x");
-    });
-  });
-
-  describe("MsBridgeService construction", () => {
-    test("creates instance with config", () => {
-      const service = newService();
-      expect(service).toBeTruthy();
-      expect(service.store).toBeTruthy();
-      expect(service.callbacks).toBeTruthy();
-    });
-
-    test("throws if logger is missing", () => {
-      expect(
-        () =>
-          new MsBridgeService(makeConfig(), {
-            tracer: createMockTracer(),
-            clock: createMockClock(),
-            discussionClient: createMockDiscussionClient(),
-            adapter: makeAdapter(),
-          }),
-      ).toThrow("logger is required");
-    });
-
-    test("throws if tracer is missing", () => {
-      expect(
-        () =>
-          new MsBridgeService(makeConfig(), {
-            logger: createMockLogger(),
-            discussionClient: createMockDiscussionClient(),
-            adapter: makeAdapter(),
-          }),
-      ).toThrow("tracer is required");
-    });
-
-    test("throws if storage is missing", () => {
-      expect(
-        () =>
-          new MsBridgeService(makeConfig(), {
-            logger: createMockLogger(),
-            tracer: createMockTracer(),
-            clock: createMockClock(),
-            adapter: makeAdapter(),
-          }),
-      ).toThrow("discussionClient is required");
-    });
-  });
-
   describe("callback handler", () => {
     let service;
     let adapter;
